+++ test/unit/executor_sampler/test_sampler_v2_options.py
@@ -23,84 +23,6 @@ from qiskit_ibm_runtime.options_models.sampler_options import SamplerOptions
 from ...ibm_test_case import IBMTestCase
 
 
-class TestSamplerOptionsToExecutorOptions(IBMTestCase):
-    """Tests for SamplerOptions.to_executor_options() method."""
-
-    def test_default_options_mapping(self):
-        """Test that default options are correctly mapped."""
-        options = SamplerOptions()
-        executor_options = options.to_executor_options()
-
-        # Check default execution options
-        self.assertEqual(executor_options.execution.init_qubits, True)
-        self.assertIsNone(executor_options.execution.rep_delay)
-
-        # Check default environment options
-        self.assertEqual(executor_options.environment.log_level, "WARNING")
-        self.assertEqual(executor_options.environment.job_tags, [])
-        self.assertEqual(executor_options.environment.private, False)
-        self.assertIsNone(executor_options.environment.max_execution_time)
-        self.assertIsNone(executor_options.environment.image)
-
-    def test_experimental_image_not_set(self):
-        """Test that image is None when experimental is empty."""
-        options = SamplerOptions()
-        options.experimental = {}
-        executor_options = options.to_executor_options()
-
-        self.assertIsNone(executor_options.environment.image)
-
-    def test_experimental_other_keys_ignored(self):
-        """Test that other experimental keys don't affect mapping."""
-        options = SamplerOptions()
-        options.experimental = {"image": "test:v1", "other_key": "value"}
-        executor_options = options.to_executor_options()
-
-        # Only image should be mapped
-        self.assertEqual(executor_options.environment.image, "test:v1")
-
-    def test_all_options_mapping(self):
-        """Test mapping of all supported options together."""
-        options = SamplerOptions()
-        options.execution.init_qubits = False
-        options.execution.rep_delay = 0.0002
-        options.environment.log_level = "INFO"
-        options.environment.job_tags = ["test1", "test2"]
-        options.environment.private = True
-        options.max_execution_time = 300
-        options.experimental = {"image": "test-image:latest"}
-
-        executor_options = options.to_executor_options()
-
-        self.assertEqual(executor_options.execution.init_qubits, False)
-        self.assertEqual(executor_options.execution.rep_delay, 0.0002)
-        self.assertEqual(executor_options.environment.log_level, "INFO")
-        self.assertEqual(executor_options.environment.job_tags, ["test1", "test2"])
-        self.assertEqual(executor_options.environment.private, True)
-        self.assertEqual(executor_options.environment.max_execution_time, 300)
-        self.assertEqual(executor_options.environment.image, "test-image:latest")
-
-    def test_experimental_dict_carry_over(self):
-        """Test that experimental dict is carried over to executor options."""
-        options = SamplerOptions()
-        options.experimental = {"custom_key": "custom_value", "another_key": 123}
-        executor_options = options.to_executor_options()
-
-        # Check that experimental dict is carried over
-        self.assertEqual(executor_options.experimental["custom_key"], "custom_value")
-        self.assertEqual(executor_options.experimental["another_key"], 123)
-
-    def test_experimental_dict_execution_mapping(self):
-        """Execution-related entries in `experimental.execution` must map to executor options."""
-        options = SamplerOptions()
-        options.experimental = {"execution": {"stretch_values": True, "scheduler_timing": True}}
-        executor_options = options.to_executor_options()
-
-        # Check that experimental dict options map to executor options.
-        self.assertEqual(executor_options.execution.stretch_values, True)
-        self.assertEqual(executor_options.execution.scheduler_timing, True)
-
-
 class TestSamplerUsingOptions(IBMTestCase):
     """Tests option setting on the ``Sampler`` class."""
