+++ test/unit/options_models/test_converters.py
@@ -12,12 +12,8 @@
 
 """Tests the converters for options models."""
 
-from qiskit_ibm_runtime.options_models.converters import (
-    estimator_options_to_executor_options,
-    sampler_option_to_executor_options,
-)
-from qiskit_ibm_runtime.options_models.estimator import EstimatorOptions
-from qiskit_ibm_runtime.options_models.sampler import SamplerOptions
+from qiskit_ibm_runtime.options_models.estimator_options import EstimatorOptions
+from qiskit_ibm_runtime.options_models.sampler_options import SamplerOptions
 
 from ...ibm_test_case import IBMTestCase
 
@@ -28,7 +24,7 @@ class TestSamplerOptionsToExecutorOptions(IBMTestCase):
     def test_default_options_mapping(self):
         """Test that default options are correctly mapped."""
         options = SamplerOptions()
-        executor_options = sampler_option_to_executor_options(options)
+        executor_options = options.to_executor_options()
 
         # Check default execution options
         self.assertEqual(executor_options.execution.init_qubits, True)
@@ -52,7 +48,7 @@ class TestSamplerOptionsToExecutorOptions(IBMTestCase):
         options.max_execution_time = 300
         options.experimental = {"image": "test-image:latest"}
 
-        executor_options = sampler_option_to_executor_options(options)
+        executor_options = options.to_executor_options()
 
         self.assertEqual(executor_options.execution.init_qubits, False)
         self.assertEqual(executor_options.execution.rep_delay, 0.0002)
@@ -66,19 +62,19 @@ class TestSamplerOptionsToExecutorOptions(IBMTestCase):
         """Test that image is None when experimental is empty."""
         options = SamplerOptions()
         options.experimental = {}
-        executor_options = sampler_option_to_executor_options(options)
+        executor_options = options.to_executor_options()
 
         self.assertIsNone(executor_options.environment.image)
 
-    def test_experimental_dict_carry_over(self):
-        """Test that experimental dict is carried over to executor options."""
+    def test_experimental_dict_carried_and_mapped(self):
+        """Test that experimental dict is carried over and mapped to executor options."""
         options = SamplerOptions()
         options.experimental = {
             "custom_key": 123,
             "image": "test:v1",
             "execution": {"stretch_values": True, "scheduler_timing": True},
         }
-        executor_options = sampler_option_to_executor_options(options)
+        executor_options = options.to_executor_options()
 
         # Experimental dict should be carried over.
         self.assertEqual(options.experimental, executor_options.experimental)
@@ -101,7 +97,7 @@ class TestEstimatorOptionsToExecutorOptions(IBMTestCase):
         options.execution.init_qubits = True
         options.execution.rep_delay = 0.001
 
-        executor_options = estimator_options_to_executor_options(options)
+        executor_options = options.to_executor_options()
 
         self.assertTrue(executor_options.execution.init_qubits)
         self.assertEqual(executor_options.execution.rep_delay, 0.001)
@@ -109,10 +105,9 @@ class TestEstimatorOptionsToExecutorOptions(IBMTestCase):
 
     def test_to_executor_options_with_experimental(self):
         """Test conversion with experimental options."""
-        options = EstimatorOptions()
-        options.experimental = {"image": "custom:image", "other": "value"}
+        options = EstimatorOptions(experimental={"image": "custom:image", "other": "value"})
 
-        executor_options = estimator_options_to_executor_options(options)
+        executor_options = options.to_executor_options()
 
         self.assertEqual(executor_options.environment.image, "custom:image")
-        self.assertEqual(executor_options.experimental, options.experimental)
+        self.assertEqual(executor_options.experimental.get("other"), "value")
