+++ test/unit/options_models/test_sampler.py
@@ -18,8 +18,10 @@ from ddt import data, ddt
 from pydantic import ValidationError
 from qiskit.transpiler import CouplingMap
 
-from qiskit_ibm_runtime.options_models.sampler import SamplerOptions
-from qiskit_ibm_runtime.options_models.simulator import SimulatorOptions
+from qiskit_ibm_runtime.executor_sampler import SamplerV2
+from qiskit_ibm_runtime.fake_provider import FakeBrisbane
+from qiskit_ibm_runtime.options_models.sampler_options import SamplerOptions
+from qiskit_ibm_runtime.options_models.simulator_options import SimulatorOptions
 
 from ...ibm_test_case import IBMTestCase
 
@@ -37,13 +39,36 @@ class TestSimulatorOptions(IBMTestCase):
         self.assertIsNone(options.simulator.coupling_map)
         self.assertIsNone(options.simulator.basis_gates)
 
-    @data([[0, 1], [1, 2]], CouplingMap([[0, 1], [1, 2]]))
-    def test_coupling_map_valid(self, coupling_map):
-        """Test setting coupling map."""
+    def test_simulator_options_set_coupling_map(self):
+        """Test setting coupling map as list."""
         options = SamplerOptions()
+        coupling_map = [[0, 1], [1, 2], [2, 3]]
         options.simulator.coupling_map = coupling_map
 
-        self.assertEqual(options.simulator.coupling_map, coupling_map)
+        self.assertEqual(options.simulator.coupling_map, CouplingMap(coupling_map))
+
+    def test_simulator_options_set_coupling_map_from_qiskit(self):
+        """Test setting coupling map from Qiskit CouplingMap object."""
+        options = SamplerOptions()
+        qiskit_coupling_map = CouplingMap([[0, 1], [1, 2]])
+        options.simulator.coupling_map = qiskit_coupling_map
+
+        self.assertEqual(options.simulator.coupling_map, qiskit_coupling_map)
+
+    def test_simulator_options_from_dict(self):
+        """Test constructing simulator options from dict."""
+        opts_dict = {
+            "simulator": {
+                "seed_simulator": 123,
+                "basis_gates": ["h", "cx", "rz"],
+                "coupling_map": [[0, 1], [1, 2]],
+            }
+        }
+        sampler = SamplerV2(mode=FakeBrisbane(), options=opts_dict)
+
+        self.assertEqual(sampler.options.simulator.seed_simulator, 123)
+        self.assertEqual(sampler.options.simulator.basis_gates, ["h", "cx", "rz"])
+        self.assertEqual(sampler.options.simulator.coupling_map, CouplingMap([[0, 1], [1, 2]]))
 
     @data("bad_input", [1, 2, 3], [[0, 1], [-1, 0]])
     def test_coupling_map_invalid_type_raises(self, input):
@@ -53,7 +78,7 @@ class TestSimulatorOptions(IBMTestCase):
 
     def test_noise_model_invalid_type_no_aer_raises(self):
         """Passing a non-dict noise_model raises when Aer is not installed."""
-        with patch("qiskit_ibm_runtime.options_models.simulator.optionals.HAS_AER", False):
+        with patch("qiskit_ibm_runtime.options_models.simulator_options.optionals.HAS_AER", False):
             with self.assertRaises(ValidationError):
                 SimulatorOptions(noise_model=object())
